feat: show etp for multiple hospitals
